Reorganize file and component structure.
